Add chancges based on comments from @artem-trubin. Add favicon. (^-^)
It would be better to just store the data from the first call
to the JSON somewhere and use it, instead of reaching out
to this function every time someone changes a tab.
https://github.com/kottans/frontend-2022-homeworks/pull/145#discussion_r950833991
Resolved

Also, you already have the same string in the main-script.js file.
Maybe consider moving it to the readJSON.js file. Though,
if you remove readJSON function usage here, maybe you don't need
readJSON.js file at all and you can keep this function
in the main-script.js.
https://github.com/kottans/frontend-2022-homeworks/pull/145#discussion_r950834145
Resolved

That's a lot of reaching out to DOM just to switch content on one action.
Consider wrapping all these elements with a document fragment and using only one appendChild.
https://github.com/kottans/frontend-2022-homeworks/pull/145#discussion_r950834232
Resolved

This looks a little confusing. The purpose of this function is to render
an AsideMenu. When you name function makeAsideMenu it usually means that
it should return some kind of element, and this function does do that,
but you're not using this element in any way, because what it also does
is render the element on line 23 right away. Either rename the function
to something like renderAsideMenu and remove the return or keep the return,
remove rendering here and render somewhere outside this function using
the returned value.
https://github.com/kottans/frontend-2022-homeworks/pull/145#discussion_r950835025
Resolved